Added constants and checking whether file with same name already exists.

+++ fileuploadservice/src/main/java/com/antiplagiarism/fileuploadservice/core/services/FileStorageService.java
@@ -7,7 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -35,14 +34,19 @@ public class FileStorageService implements IStorageService {
         try {
 
             if (fileName.contains("..")) {
-                throw new IllegalStateException("Sorry! Filename contains invalid path sequence " + fileName);
+                throw new IllegalStateException(FileExceptions.SORRY_FILENAME_CONTAINS_INVALID_PATH_SEQUENCE.getValue() + fileName);
             }
 
             Path targetLocation = this.fileStorageLocation.resolve(fileName);
+
+            if(targetLocation.toFile().exists()){
+                throw new IllegalStateException(FileExceptions.FILE_WITH_THAT_NAME_ALREADY_EXISTS.getValue());
+            }
+
             Files.copy(multipartFile.getInputStream(), targetLocation, StandardCopyOption.REPLACE_EXISTING);
 
         } catch (IOException ex) {
-            throw new IllegalStateException("Could not store file " + fileName + ". Please try again!", ex);
+            throw new IllegalStateException(FileExceptions.COULD_NOT_STORE_FILE.getValue() + fileName + FileExceptions.PLEASE_TRY_AGAIN.getValue(), ex);
         }
     }
 }
